Исправлен Header в форме Входа и регистрации
