+++ src/components/LogoOnlyHeader.jsx
@@ -1,0 +1,79 @@
+import React from "react";
+
+const LogoOnlyHeader = () => {
+  return (
+    <header
+      style={{
+        display: "flex",
+        justifyContent: "space-between",
+        alignItems: "center",
+        padding: "20px 40px",
+        backgroundColor: "#fff",
+        borderBottom: "1px solid #e0e0e0",
+        boxShadow: "0 2px 4px rgba(0, 0, 0, 0.05)",
+      }}
+    >
+      <div style={{ display: "flex", alignItems: "center" }}>
+        <svg
+          width="144"
+          height="19"
+          viewBox="0 0 144 19"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <path
+            d="M13.4872 0H5.52225C2.06253 0 0 2.0615 0 5.5195V13.471C0 16.9385 2.06253 19 5.52225 19H13.4777C16.9374 19 19 16.9385 19 13.4805V5.5195C19.0095 2.0615 16.9469 0 13.4872 0ZM6.75787 14.155C6.75787 14.421 6.54876 14.63 6.28263 14.63H3.63081C3.36468 14.63 3.15557 14.421 3.15557 14.155V9.766C3.15557 9.1675 3.64031 8.683 4.23911 8.683H6.28263C6.54876 8.683 6.75787 8.892 6.75787 9.158V14.155ZM11.3011 14.155C11.3011 14.421 11.092 14.63 10.8259 14.63H8.17407C7.90794 14.63 7.69884 14.421 7.69884 14.155V5.453C7.69884 4.8545 8.18358 4.37 8.78238 4.37H10.2271C10.8259 4.37 11.3106 4.8545 11.3106 5.453V14.155H11.3011ZM15.8539 14.155C15.8539 14.421 15.6448 14.63 15.3787 14.63H12.7268C12.4607 14.63 12.2516 14.421 12.2516 14.155V10.7825C12.2516 10.5165 12.4607 10.3075 12.7268 10.3075H14.7704C15.3692 10.3075 15.8539 10.792 15.8539 11.3905V14.155Z"
+            fill="black"
+          />
+          <path
+            d="M33.0783 15.192C32.193 15.192 31.3343 15.0747 30.5023 14.84C29.681 14.5947 29.0196 14.28 28.5183 13.896L29.3983 11.944C29.8783 12.296 30.449 12.5787 31.1103 12.792C31.7716 13.0053 32.433 13.112 33.0943 13.112C33.8303 13.112 34.3743 13.0053 34.7263 12.792C35.0783 12.568 35.2543 12.2747 35.2543 11.912C35.2543 11.6453 35.1476 11.4267 34.9343 11.256C34.7316 11.0747 34.465 10.9307 34.1343 10.824C33.8143 10.7173 33.377 10.6 32.8223 10.472C31.969 10.2693 31.2703 10.0667 30.7263 9.864C30.1823 9.66133 29.713 9.336 29.3183 8.888C28.9343 8.44 28.7423 7.84267 28.7423 7.096C28.7423 6.44533 28.9183 5.85867 29.2703 5.336C29.6223 4.80267 30.1503 4.38133 30.8543 4.072C31.569 3.76267 32.4383 3.608 33.4623 3.608C34.177 3.608 34.8756 3.69333 35.5583 3.864C36.241 4.03467 36.8383 4.28 37.3503 4.6L36.5503 6.568C35.5156 5.98133 34.481 5.688 33.4463 5.688C32.721 5.688 32.1823 5.80533 31.8303 6.04C31.489 6.27467 31.3183 6.584 31.3183 6.968C31.3183 7.352 31.5156 7.64 31.9103 7.832C32.3156 8.01333 32.929 8.19466 33.7503 8.376C34.6036 8.57867 35.3023 8.78133 35.8463 8.984C36.3903 9.18667 36.8543 9.50667 37.2383 9.944C37.633 10.3813 37.8303 10.9733 37.8303 11.72C37.8303 12.36 37.649 12.9467 37.2863 13.48C36.9343 14.0027 36.401 14.4187 35.6863 14.728C34.9716 15.0373 34.1023 15.192 33.0783 15.192Z"
+            fill="black"
+          />
+          <path
+            d="M43.1214 11.624L41.9214 12.808V15H39.4254V3.128H41.9214V9.848L45.5694 6.392H48.5454L44.9614 10.04L48.8654 15H45.8414L43.1214 11.624Z"
+            fill="black"
+          />
+          <path
+            d="M58.0319 6.392L54.1439 15.528C53.7493 16.52 53.2586 17.2187 52.6719 17.624C52.0959 18.0293 51.3973 18.232 50.5759 18.232C50.1279 18.232 49.6853 18.1627 49.2479 18.024C48.8106 17.8853 48.4533 17.6933 48.1759 17.448L49.0879 15.672C49.2799 15.8427 49.4986 15.976 49.7439 16.072C49.9999 16.168 50.2506 16.216 50.4959 16.216C50.8373 16.216 51.1146 16.1307 51.3279 15.96C51.5413 15.8 51.7333 15.528 51.9039 15.144L51.9359 15.064L48.2079 6.392H50.7839L53.1999 12.232L55.6319 6.392H58.0319Z"
+            fill="black"
+          />
+          <path
+            d="M64.1384 6.264C64.9384 6.264 65.6638 6.45067 66.3144 6.824C66.9758 7.18667 67.4931 7.704 67.8664 8.376C68.2398 9.03733 68.4264 9.81067 68.4264 10.696C68.4264 11.5813 68.2398 12.36 67.8664 13.032C67.4931 13.6933 66.9758 14.2107 66.3144 14.584C65.6638 14.9467 64.9384 15.128 64.1384 15.128C63.0398 15.128 62.1758 14.7813 61.5464 14.088V18.104H59.0504V6.392H61.4344V7.384C62.0531 6.63733 62.9544 6.264 64.1384 6.264ZM63.7064 13.08C64.3464 13.08 64.8691 12.8667 65.2744 12.44C65.6904 12.0027 65.8984 11.4213 65.8984 10.696C65.8984 9.97067 65.6904 9.39466 65.2744 8.968C64.8691 8.53067 64.3464 8.312 63.7064 8.312C63.0664 8.312 62.5384 8.53067 62.1224 8.968C61.7171 9.39466 61.5144 9.97067 61.5144 10.696C61.5144 11.4213 61.7171 12.0027 62.1224 12.44C62.5384 12.8667 63.0664 13.08 63.7064 13.08Z"
+            fill="black"
+          />
+          <path
+            d="M72.4813 7.528C72.78 7.112 73.18 6.79733 73.6813 6.584C74.1933 6.37067 74.78 6.264 75.4413 6.264V8.568C75.164 8.54667 74.9773 8.536 74.8813 8.536C74.1666 8.536 73.6066 8.73867 73.2013 9.144C72.796 9.53867 72.5933 10.136 72.5933 10.936V15H70.0973V6.392H72.4813V7.528Z"
+            fill="black"
+          />
+          <path
+            d="M80.9433 15.128C80.0366 15.128 79.2206 14.9413 78.4953 14.568C77.7806 14.184 77.2206 13.656 76.8153 12.984C76.41 12.312 76.2073 11.5493 76.2073 10.696C76.2073 9.84267 76.41 9.08 76.8153 8.408C77.2206 7.736 77.7806 7.21333 78.4953 6.84C79.2206 6.456 80.0366 6.264 80.9433 6.264C81.85 6.264 82.6606 6.456 83.3753 6.84C84.09 7.21333 84.65 7.736 85.0553 8.408C85.4606 9.08 85.6633 9.84267 85.6633 10.696C85.6633 11.5493 85.4606 12.312 85.0553 12.984C84.65 13.656 84.09 14.184 83.3753 14.568C82.6606 14.9413 81.85 15.128 80.9433 15.128ZM80.9433 13.08C81.5833 13.08 82.106 12.8667 82.5113 12.44C82.9273 12.0027 83.1353 11.4213 83.1353 10.696C83.1353 9.97067 82.9273 9.39466 82.5113 8.968C82.106 8.53067 81.5833 8.312 80.9433 8.312C80.3033 8.312 79.7753 8.53067 79.3593 8.968C78.9433 9.39466 78.7353 9.97067 78.7353 10.696C78.7353 11.4213 78.9433 12.0027 79.3593 12.44C79.7753 12.8667 80.3033 13.08 80.9433 13.08Z"
+            fill="black"
+          />
+          <path
+            d="M88.1663 15.128C87.729 15.128 87.361 14.9787 87.0623 14.68C86.7636 14.3813 86.6143 14.008 86.6143 13.56C86.6143 13.1013 86.7636 12.7333 87.0623 12.456C87.361 12.168 87.729 12.024 88.1663 12.024C88.6036 12.024 88.9716 12.168 89.2703 12.456C89.569 12.7333 89.7183 13.1013 89.7183 13.56C89.7183 14.008 89.569 14.3813 89.2703 14.68C88.9716 14.9787 88.6036 15.128 88.1663 15.128Z"
+            fill="black"
+          />
+          <path
+            d="M107.904 3.8L104.24 15H101.456L98.9924 7.416L96.4484 15H93.6804L90.0004 3.8H92.6884L95.2164 11.672L97.8564 3.8H100.256L102.816 11.736L105.424 3.8H107.904Z"
+            fill="black"
+          />
+          <path
+            d="M112.007 6.264C113.340 6.264 114.364 6.584 115.079 7.224C115.793 7.85333 116.151 8.808 116.151 10.088V15H113.815V13.928C113.345 14.728 112.471 15.128 111.191 15.128C110.529 15.128 109.953 15.016 109.463 14.792C108.983 14.568 108.615 14.2587 108.359 13.864C108.103 13.4693 107.975 13.0213 107.975 12.52C107.975 11.72 108.273 11.0907 108.871 10.632C109.479 10.1733 110.412 9.944 111.671 9.944H113.655C113.655 9.4 113.489 8.984 113.159 8.696C112.828 8.39733 112.332 8.248 111.671 8.248C111.212 8.248 110.759 8.32267 110.311 8.472C109.873 8.61067 109.5 8.80266 109.191 9.048L108.295 7.304C108.764 6.97333 109.324 6.71733 109.975 6.536C110.636 6.35467 111.313 6.264 112.007 6.264ZM111.815 13.448C112.241 13.448 112.62 13.352 112.951 13.16C113.281 12.9573 113.516 12.664 113.655 12.28V11.4H111.943C110.919 11.4 110.407 11.736 110.407 12.408C110.407 12.728 110.529 12.984 110.775 13.176C111.031 13.3573 111.377 13.448 111.815 13.448Z"
+            fill="black"
+          />
+          <path d="M118.41 3.128H120.906V15H118.41V3.128Z" fill="black" />
+          <path d="M123.222 3.128H125.718V15H123.222V3.128Z" fill="black" />
+          <path
+            d="M136.467 10.728C136.467 10.76 136.451 10.984 136.419 11.4H129.907C130.024 11.9333 130.301 12.3547 130.739 12.664C131.176 12.9733 131.72 13.128 132.371 13.128C132.819 13.128 133.213 13.064 133.555 12.936C133.907 12.7973 134.232 12.584 134.531 12.296L135.859 13.736C135.048 14.664 133.864 15.128 132.307 15.128C131.336 15.128 130.477 14.9413 129.731 14.568C128.984 14.184 128.408 13.656 128.003 12.984C127.597 12.312 127.395 11.5493 127.395 10.696C127.395 9.85333 127.592 9.096 127.987 8.424C128.392 7.74133 128.941 7.21333 129.635 6.84C130.339 6.456 131.123 6.264 131.987 6.264C132.829 6.264 133.592 6.44533 134.275 6.808C134.957 7.17067 135.491 7.69333 135.875 8.376C136.269 9.048 136.467 9.832 136.467 10.728ZM132.003 8.152C131.437 8.152 130.963 8.312 130.579 8.632C130.195 8.952 129.96 9.38933 129.875 9.944H134.115C134.029 9.4 133.795 8.968 133.411 8.648C133.027 8.31733 132.557 8.152 132.003 8.152Z"
+            fill="black"
+          />
+          <path
+            d="M143.681 14.584C143.435 14.7653 143.131 14.904 142.769 15C142.417 15.0853 142.043 15.128 141.649 15.128C140.625 15.128 139.83 14.8667 139.265 14.344C138.71 13.8213 138.433 13.0533 138.433 12.04V8.504H137.105V6.584H138.433V4.488H140.929V6.584H143.073V8.504H140.929V12.008C140.929 12.3707 141.019 12.6533 141.201 12.856C141.393 13.048 141.659 13.144 142.001 13.144C142.395 13.144 142.731 13.0373 143.009 12.824L143.681 14.584Z"
+            fill="black"
+          />
+        </svg>
+      </div>
+    </header>
+  );
+};
+export default LogoOnlyHeader;
